+++ src/PluginManager/src/Flight/FlightImportPluginBase.cpp
@@ -118,10 +118,10 @@ bool FlightImportPluginBase::importFlights(Flight &flight) noexcept
             qDebug() << QFileInfo(selectedPath).fileName() << "import" << (ok ? "SUCCESS" : "FAIL") << "in" << timer.elapsed() <<  "ms";
 #endif
             if (!ok && !baseSettings.isImportDirectoryEnabled()) {
-                QMessageBox::critical(PluginBase::getParentWidget(), tr("Import error"), tr("The file %1 could not be imported.").arg(selectedPath));
+                QMessageBox::critical(PluginBase::getParentWidget(), tr("Import Error"), tr("The file %1 could not be imported.").arg(selectedPath));
             }
         } else {
-            QMessageBox::critical(PluginBase::getParentWidget(), tr("Import error"),
+            QMessageBox::critical(PluginBase::getParentWidget(), tr("Import Error"),
                                  tr("The selected aircraft '%1' is not a known aircraft in the logbook. "
                                     "Check for spelling errors or record a flight with this aircraft first.").arg(d->selectedAircraftType.type));
         }
@@ -401,7 +401,7 @@ void FlightImportPluginBase::confirmImportError(const QString &filePath, bool &i
     messageBox->setIcon(QMessageBox::Critical);
     QPushButton *proceedButton = messageBox->addButton(tr("&Proceed"), QMessageBox::AcceptRole);
     QPushButton *ignoreAllButton = messageBox->addButton(tr("&Ignore All Failures"), QMessageBox::YesRole);
-    messageBox->setWindowTitle(tr("Import Failure"));
+    messageBox->setWindowTitle(tr("Import Error"));
     messageBox->setText(tr("The file %1 could not be imported. Do you want to proceed with the remaining files in directory %2?").arg(fileInfo.fileName(), fileInfo.dir().dirName()));
     messageBox->setInformativeText(tr("Aborting will keep the already successfully imported flights and aircraft."));
     messageBox->setStandardButtons(QMessageBox::Cancel);